move addon code to addon folder
